+++ src/deluge/io/midi/cable_types/din.h
@@ -31,5 +31,5 @@ public:
 	void sendMessage(uint8_t statusType, uint8_t channel, uint8_t data1, uint8_t data2) override;
 
 	void sendSysex(const uint8_t* data, int32_t len) override;
-	int32_t sendBufferSpace() override;
+	size_t sendBufferSpace() override;
 };
